move warehouse to commands.warehouse
